levels/06-template-based-parsing/requirements.txt
pymupdf>=1.24.0
openai>=1.0.0
python-dotenv>=1.0.0
